added new api and player packages in order to keep the code clean
